Refactor charity and donor account creation
- Added basic navbar to make testing easier
- Moved account registration to register router
