+++ views/newDonor.hbs
@@ -1,5 +1,5 @@
-<h2>Register For a New Charity Account:</h2>
-<form action="newCharity" method="POST" >
+<h2>Register For a New Donor Account:</h2>
+<form action="newDonor" method="POST" >
   <div class="form-group">
     <label for="username">Username</label>
     <input type="text" class="form-control" id="username" name="username" placeholder="">
