+++ views/register.hbs
@@ -1,32 +1,2 @@
-<h2>Register For An Account:</h2>
-<form action="/register" method="POST" >
-  <div class="form-group">
-    <label for="username">Username</label>
-    <input type="text" class="form-control" id="username" name="username" placeholder="">
-  </div>
-  <div class="form-group">
-    <label for="password">Password</label>
-    <input type="password" class="form-control" id="password" name="password" placeholder="">
-  </div>
-  <div class="form-group">
-    <label for="email">Email</label>
-    <input type="email" class="form-control" id="email" name="email" placeholder="">
-  </div>
-  <div class="form-group">
-    <label for="street">Street Address</label>
-    <input type="text" class="form-control" id="street" name="street" placeholder="">
-  </div>
-  <div class="form-group">
-    <label for="city">City</label>
-    <input type="text" class="form-control" id="city" name="city" placeholder="">
-  </div>
-  <div class="form-group">
-    <label for="state">State</label>
-    <input type="text" class="form-control" id="state" name="state" placeholder="">
-  </div>
-  <div class="form-group">
-    <label for="zipcode">Zip</label>
-    <input type="text" class="form-control" id="zipcode" name="zipcode" placeholder="">
-  </div>
-  <input type="submit" class="btn btn-danger" value="Submit">
-</form>
+<a href='register/newCharity'>I am a Charity</a>
+<a href='register/newDonor'>I am a Donor</a>
